Implement alternate indexer implementation based on libIndexStore
THIS IS A VERY ROUGH PROTOTYPE

This depends on the chain of patch in clang, starting at:
https://reviews.llvm.org/D39050
To get those, use this branch:
https://github.com/MarkZ3/clang/tree/IndexStore  (f6fa7bcf9)

And the matching revision for LLVM repo is 337115 (i.e. known to compile)

With this patch, there are two indexers: one using libIndexStore and the
"homegrown" one. To use the libIndexStore one, build with
cmake -DENABLE_LIB_INDEX_STORE=1 (... and other flags as usual)